Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "tensorflow>=2.21.0",
]
